--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -9,13 +9,7 @@ import (
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string
 }
 
 type DomainStat map[string]int
@@ -23,7 +17,6 @@ type DomainStat map[string]int
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	domainStat := make(DomainStat)
 	domain = "." + domain
-
 	scanner := bufio.NewScanner(r)
 	var user User
 	for scanner.Scan() {