--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -1,6 +1,3 @@
-//go:build bench
-// +build bench
-
 package hw10programoptimization
 
 import (
@@ -434,4 +431,21 @@ var expectedBizStat = DomainStat{
 	"zoovu.biz":         38,
 	"zooxo.biz":         33,
 	"zoozzy.biz":        23,
+}
+
+func BenchmarkGetDomainStat(b *testing.B) {
+	r, err := zip.OpenReader("testdata/users.dat.zip")
+	require.NoError(b, err)
+	defer r.Close()
+
+	require.Equal(b, 1, len(r.File))
+
+	data, err := r.File[0].Open()
+	require.NoError(b, err)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, err := GetDomainStat(data, "biz")
+		require.NoError(b, err)
+	}
 }