@khaydarov 2024-02-20
```khaydarovm@MacBook-Pro-Murod hw10_program_optimization % benchstat old.txt new.txt
goos: darwin
goarch: amd64
pkg: github.com/khaydarov/otus-golang-professional/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
                │     old.txt      │               new.txt                │
                │      sec/op      │   sec/op     vs base                 │
GetDomainStat-8   178374.603µ ± 1%   1.478µ ± 2%  -100.00% (p=0.000 n=10)
```